Ajuste cadastro de usuários
