+++ src/app/model/Restaurante.ts
@@ -1,6 +1,7 @@
 import { AbstractPojo } from "./AbstractPojo";
 import { Endereco } from "./Endereco";
 import { UsuarioRestaurante } from "./UsuarioRestaurante";
+import { Usuario } from "./Usuario";
 
 export class Restaurante extends AbstractPojo {
 
@@ -12,6 +13,13 @@ export class Restaurante extends AbstractPojo {
     telefone = null
     usuarios: UsuarioRestaurante[]
     
+    public removerUsuario(login: String) {
+        let usuarios = this.usuarios.filter((usuarioRestaurante) => {
+            return usuarioRestaurante.usuario.login == login;
+        });
+        let index = this.usuarios.indexOf(usuarios[0]);
+        this.usuarios.splice(index, 1);
+    }
 
     constructor() {
         super();
