+++ src/styles.css
@@ -4,8 +4,8 @@
 body {
     font-family: "Arial", sans-serif;  
     /*background-color: #f4f4f4;*/
-    /*, #ff3333 80%*/
-    background-color: #9b0000;
+    /*, #ff3333 80%*/    
+    background-color: #9b0000;    
     background-image: linear-gradient(to bottom, transparent, #ff6666 100%);
     background-size: auto 1280px;
 }
